feat: add check in delete command handler

+++ ms-artist/src/main/java/dev/ime/application/handler/DeleteByIdCommandHandler.java
@@ -1,21 +1,29 @@
 package dev.ime.application.handler;
 
 
+import java.util.Map;
+
 import org.springframework.stereotype.Component;
 
+import dev.ime.application.config.ApplicationConstant;
+import dev.ime.application.exception.EntityAssociatedException;
 import dev.ime.application.usecase.DeleteByIdCommand;
 import dev.ime.domain.command.Command;
 import dev.ime.domain.command.CommandHandler;
 import dev.ime.domain.port.outbound.ArtistWriteRepositoryPort;
+import dev.ime.domain.port.outbound.MediaBackupRepositoryPort;
 
 @Component
 public class DeleteByIdCommandHandler implements CommandHandler<Boolean>{
 	
 	private final ArtistWriteRepositoryPort artistWriteRepositoryPort;		
-	
-	public DeleteByIdCommandHandler(ArtistWriteRepositoryPort artistWriteRepositoryPort) {
+	private final MediaBackupRepositoryPort mediaBackupRepositoryPort;	
+
+	public DeleteByIdCommandHandler(ArtistWriteRepositoryPort artistWriteRepositoryPort,
+			MediaBackupRepositoryPort mediaBackupRepositoryPort) {
 		super();
 		this.artistWriteRepositoryPort = artistWriteRepositoryPort;
+		this.mediaBackupRepositoryPort = mediaBackupRepositoryPort;
 	}
 
 	@Override
@@ -24,6 +32,7 @@ public class DeleteByIdCommandHandler implements CommandHandler<Boolean>{
 		if (command instanceof DeleteByIdCommand deleteByIdCommand) {
 			
 			Long id = deleteByIdCommand.id();
+			if( mediaBackupRepositoryPort.existArtistId(id) ) throw new EntityAssociatedException(Map.of(ApplicationConstant.ARTISTID, String.valueOf(id)));
 			artistWriteRepositoryPort.deleteById(id);
 			return artistWriteRepositoryPort.findById(id).isEmpty();
